Refactored Experiment Code
All code has been refactored into functions, including a CLI function called Start() that will collect all relevant metadata from the user, check that all settings are correct, and run the experiment. Currently Start() is called at the end of the script so that an experiment is begun simply by running the script.